sonar-dat: move to kappa-record-db
